Changed create/edit flow for profiles
When creating a new profile, the user is now prompted for the name
via a dialog, and the profile is created as soon as the dialog
finishes. Then EditProfileActivity is started with the ID of the
new profile. This way, the edit activity always operates the same way,
and you can assume that the profile already exists within that
activity. The activity also now sets its title to the name of the
profile, and the profile can be renamed via a menu option.

Also fixed a bug where the QuickTileService crashes because there
is no selected profile yet and it tries to set it's label to that
of the selected profile. Now, if none is selected, it displays
"Select Profile".